+++ test/requirements.txt
@@ -1,6 +1,79 @@
-pytest
-boto3
-hypothesis
-moto[server]
-attrs
-cattrs
+# Direct dependencies:
+#   attrs
+#   boto3
+#   cattrs
+#   moto[server]@git+https://github.com/tjohnes/moto.git@663e321ad638d5d75f148cb484e3551cec9eec8b
+#   pytest
+annotated-types==0.5.0
+attrs==23.1.0
+aws-sam-translator==1.75.0
+aws-xray-sdk==2.12.0
+blinker==1.6.2
+boto3==1.28.55
+botocore==1.31.55
+cattrs==23.1.2
+certifi==2023.7.22
+cffi==1.15.1
+cfn-lint==0.80.3
+charset-normalizer==3.2.0
+click==8.1.7
+cryptography==41.0.4
+docker==6.1.3
+ecdsa==0.18.0
+Flask==2.3.3
+Flask-Cors==4.0.0
+graphql-core==3.2.3
+idna==3.4
+iniconfig==2.0.0
+itsdangerous==2.1.2
+Jinja2==3.1.2
+jmespath==1.0.1
+jschema-to-python==1.2.3
+jsondiff==2.0.0
+jsonpatch==1.33
+jsonpickle==3.0.2
+jsonpointer==2.4
+jsonschema==4.19.1
+jsonschema-spec==0.2.4
+jsonschema-specifications==2023.7.1
+junit-xml==1.9
+lazy-object-proxy==1.9.0
+MarkupSafe==2.1.3
+moto @ git+https://github.com/tjohnes/moto.git@663e321ad638d5d75f148cb484e3551cec9eec8b
+mpmath==1.3.0
+networkx==3.1
+openapi-schema-validator==0.6.1
+openapi-spec-validator==0.6.0
+packaging==23.1
+pathable==0.4.3
+pbr==5.11.1
+pluggy==1.3.0
+py-partiql-parser==0.3.6
+pyasn1==0.5.0
+pycparser==2.21
+pydantic==2.4.1
+pydantic_core==2.10.1
+pyparsing==3.1.1
+pytest==7.4.2
+python-dateutil==2.8.2
+python-jose==3.3.0
+PyYAML==6.0.1
+referencing==0.30.2
+regex==2023.8.8
+requests==2.31.0
+responses==0.23.3
+rfc3339-validator==0.1.4
+rpds-py==0.10.3
+rsa==4.9
+s3transfer==0.7.0
+sarif-om==1.0.4
+six==1.16.0
+sshpubkeys==3.3.1
+sympy==1.12
+types-PyYAML==6.0.12.12
+typing_extensions==4.8.0
+urllib3==1.26.16
+websocket-client==1.6.3
+Werkzeug==2.3.7
+wrapt==1.15.0
+xmltodict==0.13.0
