+++ test/terraform/node/main.tf
@@ -6,11 +6,8 @@ provider "aws" {
   }
 }
 
-
-
-
 module "node" {
-  source = "../../modules/aws/node"
+  source = "../../../modules/aws/node"
 
   wait = false
